Use git fetch --prune
I haven't fully understood under which circumstances this issue
appears (something about remote branches being renamed), but
apparently it's possible for fetches to fail if branches are not
pruned.

To compensate, change the code to not delete information for any
branches that are already in the commits file. Maybe we should
discard those if the upstream branch is dropped, but for now
retain the previous behavior.